webtoon-gallery/backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
python-multipart==0.0.20
pillow==11.1.0
pydantic==2.10.5
aiofiles==24.1.0
motor==3.6.0
python-dotenv==1.0.1